# openapi.yaml
openapi: 3.0.0
info:
  title: API E-commerce Sederhana
  description: API untuk mengelola pengguna, toko, produk, pesanan, dan transaksi.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Server Lokal
components:
  schemas:
...
    OrderStatus:
      type: string
      enum: [pending, paid, shipped, delivered, cancelled, failed]
      description: Status pesanan.
...
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int32
          readOnly: true
        store_id:
          type: integer
        category_id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        stock_quantity:
          type: integer
        is_active:
          type: boolean
    NewProduct: 
      type: object
      required: [name, price]
      properties:
        store_id:
          type: integer
          description: ID Toko (Otomatis dari konteks penjual/token)
          example: 1
        category_id:
          type: integer
          example: 5
        name:
          type: string
          example: Mouse Gaming X
        description:
          type: string
          example: Mouse ergonomis dengan sensor presisi.
        price:
          type: number
          format: float
          example: 350000.00
        stock_quantity:
          type: integer
          example: 50
        is_active:
          type: boolean
          example: true
...
    NewOrder:
      type: object
      required: [items, shipping_address]
      properties:
        shipping_address:
          type: string
          example: Jl. Sudirman No. 12, Jakarta Selatan
        items:
          type: array
          items:
            type: object
            required: [product_id, quantity]
            properties:
              product_id:
                type: integer
                example: 101
              quantity:
                type: integer
                example: 2
    Order: 
      type: object
      properties:
        id:
          type: string
          example: ORD-20251009-0001
          readOnly: true
        user_id:
          type: integer
          readOnly: true
        total_amount:
          type: number
          format: float
          readOnly: true
        status:
          $ref: '#/components/schemas/OrderStatus'
        created_at:
          type: string
          format: date-time
          readOnly: true
...
paths:
...
  /products:
    get: # ✅ GET DAFTAR DITAMBAHKAN
      tags:
        - Products
      summary: Daftar Produk
      description: Mengambil daftar semua produk dengan opsi filter/paginasi.
      parameters:
        - name: storeId
          in: query
          schema:
            type: integer
            description: Filter berdasarkan ID toko.
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Daftar produk.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags:
        - Products
      summary: Membuat Produk Baru
      description: Menambahkan produk baru ke toko penjual. (Memerlukan peran 'seller')
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProduct'
      responses:
        '201':
          description: Produk berhasil dibuat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '403':
          description: Akses ditolak.
  /products/{productId}:
    get: # ✅ GET DETAIL DITAMBAHKAN
      tags:
        - Products
      summary: Mendapatkan Detail Produk
      description: Mengambil detail satu produk berdasarkan ID.
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detail produk.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Produk tidak ditemukan.
    put:
      tags:
        - Products
      summary: Memperbarui Produk
      description: Memperbarui detail produk yang ada. (Hanya penjual pemilik toko yang bisa)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProduct'
      responses:
        '200':
          description: Produk berhasil diperbarui.
        '404':
          description: Produk tidak ditemukan.
        '403':
          description: Tidak diizinkan.
    delete: # ✅ DELETE DITAMBAHKAN
      tags:
        - Products
      summary: Menghapus Produk
      description: Menghapus produk dari toko. (Hanya pemilik atau Admin)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Produk berhasil dihapus (No Content).
        '404':
          description: Produk tidak ditemukan.

  # -----------------------------------------------------
  # --- ENDPOINT PESANAN (orders) ---
  # -----------------------------------------------------
  /orders:
    get: # ✅ GET DAFTAR DITAMBAHKAN
      tags:
        - Orders
      summary: Daftar Pesanan Saya
      description: Mengambil daftar pesanan yang dibuat oleh pengguna yang terautentikasi.
      responses:
        '200':
          description: Daftar pesanan.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Tidak terautentikasi.
    post:
      tags:
        - Orders
      summary: Membuat Pesanan Baru
      description: Pembeli membuat pesanan baru.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewOrder'
      responses:
        '201':
          description: Pesanan berhasil dibuat.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: ORD-20251009-0001
                  status:
                    $ref: '#/components/schemas/OrderStatus'
        '400':
          description: Stok tidak cukup atau input tidak valid.
...